refactor(dashboard): extract bar-shared and kpi-shared SSOT modules
Deduplicate types, constants, and pure helpers between Preact/Solid
component pairs (Bar, KpiCell, KpiStrip).

- Add bar-shared.ts with BarKind, BarProps, FILL_COLOR, barPercent
- Add kpi-shared.ts with KpiCellProps, KpiStripVariant, kpiCellAriaLabel,
  VALUE_COLOR_BY_KIND, DELTA_COLOR_BY_DIRECTION, MONO_STACK, etc.
- Redirect all imports from component files to shared modules
- Add bar.solid.test.tsx for Solid Bar DOM contract coverage
- Clean up kpi-strip-island.ts GlobalWithProcess type guard
- Fix bar.solid.tsx local BarKind import scope

Net -247 lines across 19 files. All 149 Solid tests and 58 Preact
KPI/Bar tests pass. tsc --noEmit clean. No circular dependencies.
