+++ dashboard/src/components/overview/overview.ts
@@ -12,7 +12,7 @@ import { useMemo } from 'preact/hooks'
 import { TimeAgo } from '../common/time-ago'
 import { StatusDot } from '../common/status-dot'
 import { RouteLink } from '../common/route-link'
-import type { KpiCellKind } from '../kpi-cell'
+import type { KpiCellKind } from '../kpi-shared'
 import { KpiStripIsland } from '../kpi-strip-island'
 import { AgentAvatar } from './agent-avatar'
 import { missionSnapshot } from '../../mission-store'
